# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/fee/audit": {
      "post": {
        "operationId": "FeeController_auditCompliance",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeeAuditInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeAuditOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Fee"
        ]
      }
    },
...
    "/attendance/student/{student_id}/subject-wise": {
      "get": {
        "operationId": "AttendanceController_getSubjectWiseAttendance",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "student_id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "example": "uuid-string",
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceOutputDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/forum": {
      "post": {
        "operationId": "ForumController_createPost",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Forum"
        ]
      }
    },
    "/forum/{postId}/reply": {
      "post": {
        "operationId": "ForumController_replyToPost",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Forum"
        ]
      }
    },
    "/forum/domain/{facultyId}": {
      "get": {
        "operationId": "ForumController_getDomainFilteredPosts",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facultyId",
            "required": true,
            "in": "path",
            "description": "Faculty ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostOutputDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Forum"
        ]
      }
    },
...
    "/assessments": {
      "post": {
        "operationId": "AssessmentController_createAssessment",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment"
        ]
      }
    },
    "/assessments/grade": {
      "patch": {
        "operationId": "AssessmentController_gradeAssessment",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarksEntryOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Assessment"
        ]
      }
    },
    "/courses/syllabus": {
      "patch": {
        "operationId": "AssessmentController_updateSyllabus",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "student",
                "faculty",
                "admin",
                "academic_head"
              ]
            },
            "description": "Role for Role-Based Access Control"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyllabusUpdateInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Assessment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Academic Management System API",
    "description": "In-memory NestJS backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateFeeAuditInputDto": {
        "type": "object",
        "properties": {
          "year_id": {
            "type": "string",
            "description": "UUID of the academic year",
            "example": "uuid-string"
          }
        },
        "required": [
          "year_id"
        ]
      },
      "PendingStudentDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        },
        "required": [
          "student_id",
          "amount"
        ]
      },
      "FeeAuditOutputDto": {
        "type": "object",
        "properties": {
          "year_id": {
            "type": "string"
          },
          "compliant_count": {
            "type": "number"
          },
          "non_compliant_count": {
            "type": "number"
          },
          "pending_students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingStudentDto"
            }
          }
        },
        "required": [
          "year_id",
          "compliant_count",
          "non_compliant_count",
          "pending_students"
        ]
      },
...
      "AttendanceOutputDto": {
        "type": "object",
        "properties": {
          "course_id": {
            "type": "string"
          },
          "course_name": {
            "type": "string"
          },
          "total_classes": {
            "type": "number"
          },
          "attended": {
            "type": "number"
          },
          "percentage": {
            "type": "number"
          },
          "flagged": {
            "type": "boolean"
          }
        },
        "required": [
          "course_id",
          "course_name",
          "total_classes",
          "attended",
          "percentage",
          "flagged"
        ]
      },
...
      "CreatePostInputDto": {
        "type": "object",
        "properties": {
          "topic_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "topic_id",
          "content"
        ]
      },
      "PostOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "topic_id": {
            "type": "string"
          },
          "author_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "replies_count": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "topic_id",
          "author_id",
          "content",
          "created_at",
          "replies_count"
        ]
      },
      "ReplyInputDto": {
        "type": "object",
        "properties": {
          "post_id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "post_id",
          "content"
        ]
      },
...
      "CreateAssessmentInputDto": {
        "type": "object",
        "properties": {
          "course_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "QUIZ",
              "ASSIGNMENT",
              "EXAM"
            ]
          },
          "max_marks": {
            "type": "number"
          },
          "due_date": {
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "title",
          "type",
          "max_marks",
          "due_date"
        ]
      },
      "AssessmentOutputDto": {
        "type": "object",
        "properties": {
          "assessment_id": {
            "type": "string"
          },
          "course_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "max_marks": {
            "type": "number"
          },
          "due_date": {
            "type": "string"
          }
        },
        "required": [
          "assessment_id",
          "course_id",
          "title",
          "type",
          "max_marks",
          "due_date"
        ]
      },
      "GradeInputDto": {
        "type": "object",
        "properties": {
          "entry_id": {
            "type": "string"
          },
          "marks": {
            "type": "number"
          }
        },
        "required": [
          "entry_id",
          "marks"
        ]
      },
      "MarksEntryOutputDto": {
        "type": "object",
        "properties": {
          "entry_id": {
            "type": "string"
          },
          "assessment_id": {
            "type": "string"
          },
          "student_id": {
            "type": "string"
          },
          "marks": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "entry_id",
          "assessment_id",
          "student_id",
          "marks",
          "status"
        ]
      },
      "SyllabusUpdateInputDto": {
        "type": "object",
        "properties": {
          "course_id": {
            "type": "string"
          },
          "topic": {
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic"
        ]
      },
...
    }
  }
}